Fix two type "errors"
They're both spots where we're doing something a bit tricky. One is a formally ambiguous overload (can't be helped as the actual type model seems to be impossible to express, and does mean that the rest of the code is better) and the other is a place with a weird hack to work around a fight between typing and dataclasses over whether types are evaluated during parsing or not.